Support ACI images in the Docker Engine
== Overview

This PR provides simple, functional support for the App Container Spec within
the Docker Engine. Specifically, it implements the ability to natively run App
Container Images (ACIs) and also to perform image discovery to retrieve images.
It achieves this with a minimally invasive patch which preserves the Docker
user experience.

== Examples

The following examples demonstrate basic usage of Docker with ACI support.

An ACI can be directly pulled by Docker using image discovery:
  # docker pull --format aci coreos.com/etcd:v2.0.0

Alternatively, ACIs can be imported to Docker from a file on disk or directly
from an http URL:
  # docker pull --format aci file:///tmp/aciexamples-1-linux-amd64.aci
  # docker pull --format aci http://example.com/aciexamples-1-linux-amd64.aci

Listing images, including ACI images, works as expected:
  # docker images

ACIs can be run directly; referencing an image by its name will transitively
perform image discovery if necessary:
  # docker run --format aci coreos.com/etcd

ACIs can also be referenced by their ImageID:
  # docker run --format aci fa1cb92dc276b0f9bedf87981e61ecde93cc16432d2441f23aa006a42bb873df
  # docker run --format aci fa1cb92dc276

Docker create/start is also supported:
  # docker create --format aci coreos.com/etcd
  b2eb6289ced6506316b31daac101e3f3143f11ff96d046aca3338c9b9de934dc
  # docker start -i -a b2eb6289ced6506316b31daac101e3f3143f11ff96d046aca3338c9b9de934dc

Deleting ACI images:
  # docker rmi --format aci coreos.com/etcd

== How it works

An ACI image contains a rootfs directory and a manifest file. When an ACI image
is pulled, the rootfs is converted to layer.tar. layer.tar and manifest are
then stored in the normal Docker graphstore. For example:

  # ls -l /var/lib/docker/graph/fa1cb92dc276b0f9bedf87981e61ecde93cc16432d2441f23aa006a42bb873df/
  total 12364
  -rw-r--r-- 1 root root 12653824 Feb 13 11:24 layer.tar
  -rw-r--r-- 1 root root      228 Feb 13 11:24 manifest
  #

The directory name is the sha512sum of the ACI image truncated to 64
characters.

A new entry in the image repository (/var/lib/docker/repositories-*) is
created. Example:

  "ACIRepo" : {
      "coreos.com/etcd" : "fa1cb92dc276b0f9bedf87981e61ecde93cc16432d2441f23aa006a42bb873df",
      "github.com/endocode/aciexamples" : "e3cd06d49820d9457115536ddbf2f05c4247d22190abed8f1c70551149add72f"
   }

When creating a container based on an ACI image, its configuration file
/var/lib/docker/containers/<container_id>/config.json is generated from the
manifest file and the command line parameters. For example, when the manifest
contains:

   "app": {
        "exec": [
             "/etcd"
        ],

Docker will generate config.json will have the following entrypoint:

     "Entrypoint" : [
         "/etcd"
      ],

== Known Limitations

The following are known limitations with the current state of this patch.

- **labels**: the ACI discovery mechanism defines various labels (for example:
version, os, and arch). They are not stored in the Docker repository, so two
versions of an ACI image with the same name cannot be stored at the same time.

- **isolators**: isolators specified in the manifest of an ACI are ignored.
However, options such as --cap-add, --cap-drop, --privileged should still work
when used on an ACI.

- **dependencies**: an ACI image can depend on other images in a similar way to
layers in Docker. This is not implemented yet. Only ACI images without
dependencies are supported.

- **unix group**: a manifest can specify to run the program in the container as
a specific unix group. This is not implemented yet.

- **unix user or group referring to a file**: instead of specifying a uid/gid
or username/groupname, a manifest can request to use the owner and group of an
file found at an absolute path inside the rootfs. This is not implemented yet.

- **mount points** and **ports** specified in the manifest are ignored. This is
not implemented yet.

We would really love to solve all of these limitations over time. But, before
we put in a substantial amount of work we wanted to have a wider design
proposal. So, please consider this in tandem with the issue on
http://github.com/docker/docker/issues which more broadly discusses
implementation of appc in the context of Docker.